🚧 PLAT-261 Select virtru-sdk version at runtime (#149)
To test, append `?zver=lts` to the URI, e.g. 

    https://demos.developer.virtru.com/protect-develop/?zver=lts

This will load the 1.6.3 version of the sdk, instead of the previous one. Ideally we would then be able to test against both versions in the integrated tests.

PLAT-261 - this should address this issue, when combined with the fix in https://github.com/virtru/protect-and-track/pull/146